Completed Book Api. Fixed issue with related objects (well maybe).

+++ BookStoreApi/Data/Book.cs
@@ -1,4 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
+using System.Net;
 
 namespace BookStoreApi.Data
 {
@@ -11,7 +12,7 @@ namespace BookStoreApi.Data
         public string ISBN { get; set; }
         public string Summary { get; set; }
         public string Image { get; set; }
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
         public int? AuthorId { get; set; }
 
         public virtual Author Author { get; set; }
